@@ -1,13 +1,31 @@
-﻿using Xunit;
+﻿using System.Text.Json;
+using System.Text.Json.Serialization;
+using Xunit;
 
 namespace MentorBot.Tests
 {
   public class UrlEncoderTests
   {
+    private readonly JsonSerializerOptions _opts = new ()
+    {
+      Converters =
+      {
+        new JsonStringEnumConverter(JsonNamingPolicy.CamelCase)
+      },
+      PropertyNamingPolicy = JsonNamingPolicy.CamelCase,
+    };
+
     public record SerializeObject
     {
-      public string A { get; init; }
-      public string B { get; init; }
+      public string A { get; set; }
+      public string B { get; set; }
+      public ValueThing V { get; set; }
+    }
+
+    public enum ValueThing
+    {
+      ItsA,
+      ItsB
     }
 
     [Fact]
@@ -16,11 +34,12 @@ namespace MentorBot.Tests
       var obj = new SerializeObject
       {
         A = "Taco",
-        B = "Legume"
+        B = "Legume",
+        V = ValueThing.ItsA,
       };
-      var encoded = UrlEncoder.Encode(obj);
+      var encoded = UrlEncoder.Encode(obj, _opts);
 
-      Assert.Equal("A=Taco&B=Legume", encoded);
+      Assert.Equal("a=Taco&b=Legume&v=itsA", encoded);
     }
 
     [Fact]
@@ -29,10 +48,10 @@ namespace MentorBot.Tests
       var expected = new SerializeObject
       {
         A = "Taco",
-        B = "Legume"
+        B = "Legume",
+        V = ValueThing.ItsB,
       };
-      var obj = UrlEncoder.Decode<SerializeObject>("A=Taco&B=Legume");
-
+      var obj = UrlEncoder.Decode<SerializeObject>("a=Taco&b=Legume&v=itsb", _opts);
 
       Assert.Equal(expected, obj);
     }
