@@ -1,18 +1,14 @@
 ﻿using Discord;
 using Microsoft.AspNetCore.Mvc;
-using Newtonsoft.Json;
-using Newtonsoft.Json.Converters;
 using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
 using System.Linq;
-using System.Runtime.Serialization;
 
 namespace MentorBot.Models
 {
-  public record TicketCreate
+  public class TicketCreate
   {
-    [FromQuery]
-    public string? Name { get; init; }
     [FromQuery]
     public string? UserId { get; init; }
     [FromQuery]
@@ -27,69 +23,45 @@ namespace MentorBot.Models
     public string? SessionUrl { get; init; }
     [FromQuery]
     public string? SessionWebUrl { get; init; }
-
-    public Ticket Populate(Ticket ticket)
-    {
-      return ticket with
-      {
-        Lang = Lang,
-        Desc = Desc,
-        Session = Session,
-        SessionId = SessionId,
-        SessionUrl = SessionUrl,
-        SessionWebUrl = SessionWebUrl,
-      };
-    }
   }
 
-  [JsonObject]
-  public record Ticket
+  public class Ticket
   {
-    [JsonProperty("status"), JsonConverter(typeof(StringEnumConverter))]
     public TicketStatus Status { get; set; } = TicketStatus.Requested;
+    public User User { get; set; } = new User();
+    public string? Lang { get; set; }
+    public string? Desc { get; set; }
+    public string? Session { get; set; }
+    public string? SessionId { get; set; }
+    public string? SessionUrl { get; set; }
+    public string? SessionWebUrl { get; set; }
 
-    [JsonProperty("menteeName")]
-    public string? Username { get; init; }
-    [JsonProperty("menteeId")]
-    public string UserId { get; init; } = string.Empty;
-    [JsonProperty("language")]
-    public string? Lang { get; init; }
-    [JsonProperty("description")]
-    public string? Desc { get; init; }
-    [JsonProperty("sessionName")]
-    public string? Session { get; init; }
-    [JsonProperty("sessionId")]
-    public string? SessionId { get; init; }
-    [JsonProperty("sessionUrl")]
-    public string? SessionUrl { get; init; }
-    [JsonProperty("sessionWebUrl")]
-    public string? SessionWebUrl { get; init; }
+    [Key]
+    public ulong Id { get; set; }
 
-    [JsonProperty("id")]
-    public string Id { get; init; } = string.Empty;
+    public ulong? DiscordId { get; set; }
 
-    [JsonProperty("ticket")]
-    public string LegacyTicket => Id;
+    public Mentor? Mentor { get; set; }
 
-    [JsonProperty("mentorName")]
-    public string? MentorName { get; init; } = null;
+    public DateTime Created { get; set; } = DateTime.UtcNow;
+    public DateTime? Claimed { get; set; }
+    public DateTime? Complete { get; set; }
+    public DateTime? Canceled { get; set; }
 
-    [JsonProperty("mentor")]
-    public string? LegacyMentor => MentorName;
+    public Ticket()
+    {
+    }
 
-    [JsonProperty("mentorDiscordId")]
-    public string? MentorDiscordId { get; init; } = null;
-    [JsonProperty("mentorNeosId")]
-    public string? MentorNeosId { get; init; } = null;
-
-    [JsonProperty("created")]
-    public DateTime Created { get; init; } = DateTime.UtcNow;
-    [JsonProperty("claimed")]
-    public DateTime? Claimed { get; init; } = null;
-    [JsonProperty("complete")]
-    public DateTime? Complete { get; init; } = null;
-    [JsonProperty("canceled")]
-    public DateTime? Canceled { get; init; } = null;
+    public Ticket(TicketCreate createArgs, User user)
+    {
+      User = user;
+      Lang = createArgs.Lang;
+      Desc = createArgs.Desc;
+      Session = createArgs.Session;
+      SessionId = createArgs.SessionId;
+      SessionUrl = createArgs.SessionUrl;
+      SessionWebUrl = createArgs.SessionWebUrl;
+    }
 
     private static string StatusToTitle(TicketStatus status)
     {
@@ -134,38 +106,58 @@ namespace MentorBot.Models
     {
       IEnumerable<(string Name, string Value, bool Inline)?> NullableFields()
       {
-        yield return Field("User", Username, true);
-        yield return Field("User Neos Id", UserId, true);
+        yield return Field("Ticket Number", Id.ToString());
+        yield return Field("User", User.Name, true);
+        yield return Field("User Neos Id", User.Id, true);
         yield return Field("Language", Lang);
         yield return Field("Description", Desc);
         yield return Field("Session", Session);
         yield return Field("Session ID", SessionId);
         yield return Field("Session Url", SessionUrl);
         yield return Field("Session Web Url", SessionWebUrl);
-        yield return Field("Mentor Name", MentorName, true);
-        yield return Field("Mentor Discord Link", MentorDiscordId, true);
-        yield return Field("Mentor Neos Id",
-          !string.IsNullOrWhiteSpace(MentorName) && string.IsNullOrWhiteSpace(MentorNeosId) ?
-          "<UNREGISTERED>" :
-          MentorNeosId, true);
-        yield return Field("Created", Created.ToString("u"));
-        yield return Field("Claimed", Claimed?.ToString("u"));
-        yield return Field("Completed", Complete?.ToString("u"));
-        yield return Field("Canceled", Canceled?.ToString("u"));
+        yield return Field("Mentor Name", Mentor?.Name, true);
+        yield return Field("Mentor Discord Link", Mentor?.DiscordId?.ToString(), true);
+        yield return Field("Mentor Neos ID", Mentor?.NeosId.ToString(), true);
+        yield return Field("Created", Created.ToDiscordTimecode("f"));
+        yield return Field("Claimed", Claimed?.ToDiscordTimecode("f"));
+        yield return Field("Completed", Complete?.ToDiscordTimecode("f"));
+        yield return Field("Canceled", Canceled?.ToDiscordTimecode("f"));
       }
       return NullableFields().OnlyNotNull();
     }
+
+    public TicketDto ToDto() => new(this);
+    public MentorTicketDto ToMentorDto() => new(this);
+  }
+
+  public class TicketDto {
+    protected readonly Ticket _ticket;
+    public TicketDto(Ticket ticket)
+    {
+      _ticket = ticket;
+    }
+
+    public ulong Ticket => _ticket.Id;
+    public string? Mentor => _ticket.Mentor?.Name;
+    public TicketStatus Status => _ticket.Status;
+  }
+
+  public class MentorTicketDto : TicketDto
+  {
+    public MentorTicketDto(Ticket ticket) : base(ticket)
+    {
+    }
+
+    public string? SessionId => _ticket.SessionId;
+    public string? UserId => _ticket.User.Id;
+    public string UserName => _ticket.User.Name;
   }
 
   public enum TicketStatus
   {
-    [EnumMember(Value = "requested")]
     Requested,
-    [EnumMember(Value = "responding")]
     Responding,
-    [EnumMember(Value = "completed")]
     Completed,
-    [EnumMember(Value = "canceled")]
     Canceled
   }
 
