@@ -1,5 +1,6 @@
-﻿using Microsoft.Azure.Cosmos;
-using Microsoft.Extensions.Options;
+﻿using MentorBot.Extern;
+using System;
+using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -7,45 +8,153 @@ namespace MentorBot.Models
 {
   public interface ITicketContext
   {
-    ValueTask<Ticket?> GetTicket(string id, CancellationToken cancellationToken = default);
-    ValueTask<Ticket?> CreateTicket(Ticket item, CancellationToken cancellationToken = default);
-    ValueTask<Ticket?> UpdateTicket(Ticket item, CancellationToken cancellationToken = default);
+    IQueryable<Ticket> GetIncompleteTickets();
+    ValueTask<Ticket?> GetTicketAsync(ulong ticketId, CancellationToken cancellationToken = default);
+    ValueTask<Ticket?> CreateTicketAsync(TicketCreate createArgs, CancellationToken cancellationToken = default);
+    ValueTask<Ticket?> TryCompleteTicketAsync(ulong ticketId, string mentorToken, CancellationToken cancellationToken = default);
+    ValueTask<Ticket?> TryCancelTicketAsync(ulong ticketId, CancellationToken cancellationToken = default);
+    ValueTask<Ticket?> TryClaimTicketAsync(ulong ticketId, string mentorToken, CancellationToken cancellationToken = default);
+    ValueTask<Ticket?> TryUnclaimTicketAsync(ulong ticketId, string mentorToken, CancellationToken cancellationToken = default);
+    ValueTask<Ticket?> AssignDiscordIdAsync(ulong ticket, ulong discordId, CancellationToken cancellationToken = default);
   }
 
   public class TicketContext : ITicketContext
   {
-    private readonly Container _container;
+    private readonly SignalContext _ctx;
     private readonly ITicketNotifier _notifier;
+    private readonly INeosApi _neosApi;
 
-    public TicketContext(CosmosClient dbClient, ITicketNotifier notifier, IOptionsSnapshot<CosmosOptions> options)
+    public TicketContext(SignalContext ctx, ITicketNotifier notifier, INeosApi neosApi)
     {
-      _container = dbClient.GetContainer(options.Value.Database, options.Value.TicketsContainer);
+      _ctx = ctx;
       _notifier = notifier;
+      _neosApi = neosApi;
     }
 
-    public async ValueTask<Ticket?> GetTicket(string id, CancellationToken cancellationToken = default)
+    public IQueryable<Ticket> GetIncompleteTickets()
     {
-      try
-      {
-        return await _container.ReadItemAsync<Ticket>(id.ToString(), new PartitionKey(id), cancellationToken: cancellationToken);
-      }
-      catch (CosmosException ex) when (ex.StatusCode == System.Net.HttpStatusCode.NotFound)
+      return _ctx.Tickets.Where(t => t.Status == TicketStatus.Requested || t.Status == TicketStatus.Responding);
+    }
+
+    public async ValueTask<Ticket?> GetTicketAsync(ulong ticketId, CancellationToken cancellationToken = default)
+    {
+      return await _ctx.GetTicketAsync(ticketId, cancellationToken);
+    }
+
+    public async ValueTask<Ticket?> CreateTicketAsync(TicketCreate createArgs, CancellationToken cancellationToken = default)
+    {
+      if (createArgs.UserId == null)
       {
         return null;
       }
-    }
 
-    public async ValueTask<Ticket?> CreateTicket(Ticket item, CancellationToken cancellationToken = default)
-    {
-      var ticket = await _container.CreateItemAsync(item, new PartitionKey(item.Id), cancellationToken: cancellationToken);
+      User? user = await _neosApi.GetUserAsync(createArgs.UserId, cancellationToken);
+      if (user == null)
+      {
+        return null;
+      }
+
+      Ticket ticket = new(createArgs, user)
+      {
+        Status = TicketStatus.Requested,
+        Created = DateTime.UtcNow
+      };
+
+      _ctx.Tickets.Add(ticket);
+      await _ctx.SaveChangesAsync(cancellationToken);
       _notifier.NotifyNewTicket(ticket);
       return ticket;
     }
 
-    public async ValueTask<Ticket?> UpdateTicket(Ticket item, CancellationToken cancellationToken = default)
+    public async ValueTask<Ticket?> TryClaimTicketAsync(ulong ticketId, string mentorToken, CancellationToken cancellationToken = default)
     {
-      var ticket = await _container.UpsertItemAsync(item, new PartitionKey(item.Id), cancellationToken: cancellationToken);
-      _notifier.NotifyUpdatedTicket(ticket);
+      var ticket = await _ctx.GetTicketAsync(ticketId, cancellationToken);
+      var mentor = await _ctx.GetMentorByTokenAsync(mentorToken, cancellationToken);
+      if (ticket == null || mentor == null)
+      {
+        return null;
+      }
+      if (ticket.Status == TicketStatus.Requested)
+      {
+        ticket.Mentor = mentor;
+        ticket.Status = TicketStatus.Responding;
+        ticket.Claimed = DateTime.UtcNow;
+
+        _ctx.Tickets.Update(ticket);
+        await _ctx.SaveChangesAsync(cancellationToken);
+        _notifier.NotifyUpdatedTicket(ticket);
+      }
+      return ticket;
+    }
+
+    public async ValueTask<Ticket?> TryUnclaimTicketAsync(ulong ticketId, string mentorToken, CancellationToken cancellationToken = default)
+    {
+      var ticket = await _ctx.GetTicketAsync(ticketId, cancellationToken);
+      if (ticket == null || ticket.Mentor?.Token != mentorToken)
+      {
+        return null;
+      }
+      if (ticket.Status == TicketStatus.Responding)
+      {
+        ticket.Status = TicketStatus.Requested;
+        ticket.Claimed = null;
+        ticket.Mentor = null;
+
+        _ctx.Tickets.Update(ticket);
+        await _ctx.SaveChangesAsync(cancellationToken);
+        _notifier.NotifyUpdatedTicket(ticket);
+      }
+      return ticket;
+    }
+
+    public async ValueTask<Ticket?> TryCompleteTicketAsync(ulong ticketId, string mentorToken, CancellationToken cancellationToken = default)
+    {
+      var ticket = await _ctx.GetTicketAsync(ticketId, cancellationToken);
+      if (ticket == null || ticket.Mentor?.Token != mentorToken)
+      {
+        return null;
+      }
+      if (ticket.Status == TicketStatus.Responding)
+      {
+        ticket.Status = TicketStatus.Completed;
+        ticket.Complete = DateTime.UtcNow;
+
+        _ctx.Tickets.Update(ticket);
+        await _ctx.SaveChangesAsync(cancellationToken);
+        _notifier.NotifyUpdatedTicket(ticket);
+      }
+      return ticket;
+    }
+
+    public async ValueTask<Ticket?> TryCancelTicketAsync(ulong ticketId, CancellationToken cancellationToken = default)
+    {
+      var ticket = await _ctx.GetTicketAsync(ticketId, cancellationToken);
+      if (ticket == null)
+      {
+        return null;
+      }
+      if (!ticket.Status.IsTerminal())
+      {
+        ticket.Status = TicketStatus.Canceled;
+        ticket.Canceled = DateTime.UtcNow;
+
+        _ctx.Tickets.Update(ticket);
+        await _ctx.SaveChangesAsync(cancellationToken);
+        _notifier.NotifyUpdatedTicket(ticket);
+      }
+      return ticket;
+    }
+
+    public async ValueTask<Ticket?> AssignDiscordIdAsync(ulong ticketId, ulong discordId, CancellationToken cancellationToken = default)
+    {
+      var ticket = await _ctx.GetTicketAsync(ticketId, cancellationToken);
+      if (ticket == null)
+      {
+        return null;
+      }
+      ticket.DiscordId = discordId;
+      _ctx.Tickets.Update(ticket);
+      await _ctx.SaveChangesAsync(cancellationToken);
       return ticket;
     }
   }
