@@ -1,11 +1,13 @@
-﻿using Newtonsoft.Json;
+﻿using Microsoft.EntityFrameworkCore;
+using Newtonsoft.Json;
 
 namespace MentorBot.Models
 {
+  [Owned]
   public record User
   {
     [JsonProperty("id")]
-    public string Id { get; set; } = string.Empty;
+    public string? Id { get; set; }
     [JsonProperty("name")]
     public string Name { get; set; } = string.Empty;
   }
