@@ -4,6 +4,7 @@ using System.IO;
 using System.Net.WebSockets;
 using System.Runtime.CompilerServices;
 using System.Text;
+using System.Text.Json;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -11,7 +12,7 @@ namespace MentorBot
 {
   public static class WebSocketExtensions
   {
-    public static async IAsyncEnumerable<string> ReadMessages(this WebSocket socket, [EnumeratorCancellation] CancellationToken cancelToken = default)
+    public static async IAsyncEnumerable<string> ReadRawMessages(this WebSocket socket, [EnumeratorCancellation] CancellationToken cancelToken = default)
     {
       var buffer = new byte[1024 * 4];
       WebSocketReceiveResult result;
@@ -30,7 +31,19 @@ namespace MentorBot
       } while (result.MessageType != WebSocketMessageType.Close);
     }
 
-    public static Func<string, Task> MessageSender(this WebSocket socket, CancellationToken cancellationToken = default)
+    public static async IAsyncEnumerable<T> ReadMessages<T>(this WebSocket socket, JsonSerializerOptions? serializerOptions = null, [EnumeratorCancellation] CancellationToken cancelToken = default)
+    {
+      await foreach (var message in ReadRawMessages(socket, cancelToken))
+      {
+        var resp = UrlEncoder.Decode<T>(message, serializerOptions);
+        if (resp != null)
+        {
+          yield return resp;
+        }
+      }
+    }
+
+    public static Func<string, Task> RawMessageSender(this WebSocket socket, CancellationToken cancellationToken = default)
     {
       Task t = Task.CompletedTask;
 
@@ -51,6 +64,12 @@ namespace MentorBot
 
       return msg => NextSequenced(() =>
         socket.SendAsync(Encoding.UTF8.GetBytes(msg), WebSocketMessageType.Text, true, cancellationToken));
+    }
+
+    public static Func<T, Task> MessageSender<T>(this WebSocket socket, JsonSerializerOptions? serializerOptions = null, CancellationToken cancellationToken = default)
+    {
+      var sender = RawMessageSender(socket, cancellationToken);
+      return obj => sender(obj != null ? UrlEncoder.Encode(obj, serializerOptions) : "");
     }
   }
 }
